requirements.txt
numpy==1.23.5
streamlit==1.22.0
Pillow==9.5.0
keras==2.12.0
tensorflow==2.12.0

opencv-python~=4.9.0.80
pandas~=2.0.3